Minor db package and stuff cleanup
- Moved DbResultT and * into separate object files
- Moved DBIO monad def to the top of file (this matters yeah)
- Removed caseClassToMap
- Fixed scala binary version warning
