+++ phoenix-scala/app/utils/db/package.scala
@@ -1,22 +1,53 @@
 package utils
 
-import scala.collection.generic.CanBuildFrom
 import scala.concurrent.Future
 
-import cats.data.{Validated, Xor, XorT}
+import cats.data.{Xor, XorT}
 import cats.{Applicative, Functor, Monad}
 import failures._
 import services.Result
 import slick.driver.PostgresDriver.api._
 import slick.jdbc.SQLActionBuilder
 import slick.lifted.Query
-import slick.profile.{SqlAction, SqlStreamingAction}
+import slick.profile.SqlAction
 import utils.aliases._
 import utils.time.JavaTimeSlickMapper
 
 package object db {
 
   type DbResultT[A] = XorT[DBIO, Failures, A]
+  // DBIO monad
+  implicit def dbioApplicative(implicit ec: EC): Applicative[DBIO] = new Applicative[DBIO] {
+    def ap[A, B](fa: DBIO[A])(f: DBIO[A ⇒ B]): DBIO[B] =
+      fa.flatMap(a ⇒ f.map(ff ⇒ ff(a)))
+
+    def pure[A](a: A): DBIO[A] = DBIO.successful(a)
+  }
+
+  implicit def dbioMonad(implicit ec: EC) = new Functor[DBIO] with Monad[DBIO] {
+    override def map[A, B](fa: DBIO[A])(f: A ⇒ B): DBIO[B] = fa.map(f)
+
+    override def pure[A](a: A): DBIO[A] = DBIO.successful(a)
+
+    override def flatMap[A, B](fa: DBIO[A])(f: A ⇒ DBIO[B]): DBIO[B] = fa.flatMap(f)
+  }
+
+  // implicits
+  implicit class EnrichedDbResultT[A](dbResultT: DbResultT[A]) {
+    def runTxn()(implicit db: DB): Result[A] =
+      dbResultT.value.transactionally.run()
+
+    def run()(implicit db: DB): Result[A] =
+      dbResultT.value.run()
+
+    def meh(implicit ec: EC): DbResultT[Unit] = for (_ ← * <~ dbResultT) yield {}
+  }
+
+  final implicit class EnrichedOption[A](val option: Option[A]) extends AnyVal {
+    def toXor[F](or: F): F Xor A =
+      option.fold { Xor.left[F, A](or) }(Xor.right[F, A])
+  }
+
   // Return B whenever A is inserted
   type Returning[A, B] = slick.driver.JdbcActionComponent#ReturningInsertActionComposer[A, B]
 
@@ -110,119 +141,4 @@ package object db {
     // exist.
     def safeGet(implicit ec: EC): DBIO[R] = dbio.map(_.get)
   }
-
-  // DBIO monad
-  implicit def dbioApplicative(implicit ec: EC): Applicative[DBIO] = new Applicative[DBIO] {
-    def ap[A, B](fa: DBIO[A])(f: DBIO[A ⇒ B]): DBIO[B] =
-      fa.flatMap(a ⇒ f.map(ff ⇒ ff(a)))
-
-    def pure[A](a: A): DBIO[A] = DBIO.successful(a)
-  }
-
-  implicit def dbioMonad(implicit ec: EC) = new Functor[DBIO] with Monad[DBIO] {
-    override def map[A, B](fa: DBIO[A])(f: A ⇒ B): DBIO[B] = fa.map(f)
-
-    override def pure[A](a: A): DBIO[A] = DBIO.successful(a)
-
-    override def flatMap[A, B](fa: DBIO[A])(f: A ⇒ DBIO[B]): DBIO[B] = fa.flatMap(f)
-  }
-
-  // implicits
-  implicit class EnrichedDbResultT[A](dbResultT: DbResultT[A]) {
-    def runTxn()(implicit db: DB): Result[A] =
-      dbResultT.value.transactionally.run()
-
-    def run()(implicit db: DB): Result[A] =
-      dbResultT.value.run()
-
-    def meh(implicit ec: EC): DbResultT[Unit] = for (_ ← * <~ dbResultT) yield {}
-  }
-
-  final implicit class EnrichedOption[A](val option: Option[A]) extends AnyVal {
-    def toXor[F](or: F): F Xor A =
-      option.fold { Xor.left[F, A](or) }(Xor.right[F, A])
-  }
-
-  // DbResultT
-  object DbResultT {
-
-    def apply[A](v: DBIO[Failures Xor A]): DbResultT[A] =
-      XorT[DBIO, Failures, A](v)
-
-    def pure[A](v: A)(implicit ec: EC): DbResultT[A] =
-      XorT.pure[DBIO, Failures, A](v)
-
-    def fromXor[A](v: Failures Xor A)(implicit ec: EC): DbResultT[A] =
-      v.fold(failures, good)
-
-    def fromDbio[A](v: DBIO[A])(implicit ec: EC): DbResultT[A] =
-      XorT.right[DBIO, Failures, A](v)
-
-    def fromFuture[A](v: Future[A])(implicit ec: EC): DbResultT[A] =
-      fromDbio(DBIO.from(v))
-
-    def good[A](v: A)(implicit ec: EC): DbResultT[A] =
-      XorT.right[DBIO, Failures, A](DBIO.successful(v))
-
-    def failures[A](v: Failures)(implicit ec: EC): DbResultT[A] =
-      XorT.left[DBIO, Failures, A](DBIO.successful(v))
-
-    def failure[A](f: Failure)(implicit ec: EC): DbResultT[A] =
-      failures[A](f.single)
-
-    def sequence[A, M[X] <: TraversableOnce[X]](values: M[DbResultT[A]])(
-        implicit buildFrom: CanBuildFrom[M[DbResultT[A]], A, M[A]],
-        ec: EC): DbResultT[M[A]] =
-      values
-        .foldLeft(good(buildFrom(values))) { (liftedBuilder, liftedValue) ⇒
-          for (builder ← liftedBuilder; value ← liftedValue) yield builder += value
-        }
-        .map(_.result)
-
-    def unit(implicit ec: EC): DbResultT[Unit] =
-      pure({})
-
-    def none[A](implicit ec: EC): DbResultT[Option[A]] =
-      pure(Option.empty[A])
-  }
-
-  object * {
-    def <~[A](v: DBIO[Failures Xor A]): DbResultT[A] =
-      DbResultT(v)
-
-    def <~[A](v: SqlAction[A, NoStream, Effect.All])(implicit ec: EC): DbResultT[A] =
-      DbResultT(v.map(Xor.right))
-
-    def <~[A](v: DBIO[A])(implicit ec: EC): DbResultT[A] =
-      DbResultT.fromDbio(v)
-
-    def <~[A](v: Failures Xor A)(implicit ec: EC): DbResultT[A] =
-      DbResultT.fromXor(v)
-
-    def <~[A](v: Future[Failures Xor A]): DbResultT[A] =
-      DbResultT(DBIO.from(v))
-
-    def <~[A](v: Future[A])(implicit ec: EC): DbResultT[A] =
-      DbResultT.fromFuture(v)
-
-    def <~[A](v: A)(implicit ec: EC): DbResultT[A] =
-      DbResultT.pure(v)
-
-    def <~[A](v: Validated[Failures, A])(implicit ec: EC): DbResultT[A] =
-      DbResultT.fromXor(v.toXor)
-
-    def <~[A, M[X] <: TraversableOnce[X]](v: M[DbResultT[A]])(
-        implicit buildFrom: CanBuildFrom[M[DbResultT[A]], A, M[A]],
-        ec: EC): DbResultT[M[A]] =
-      DbResultT.sequence(v)
-
-    def <~[A](v: DbResultT[A]): DbResultT[A] =
-      v
-
-    def <~[A](v: Option[DbResultT[A]])(implicit ec: EC): DbResultT[Option[A]] =
-      v.fold(DbResultT.none[A]) { dbresult ⇒
-        dbresult.map(Some(_))
-      }
-  }
-
 }
