+++ phoenix-scala/app/utils/package.scala
@@ -15,11 +15,6 @@ package object utils {
       "_" + m.group(0).toLowerCase
     })
 
-  implicit def caseClassToMap(cc: Product): Map[String, Any] = {
-    val values = cc.productIterator
-    cc.getClass.getDeclaredFields.map(_.getName → values.next).toMap
-  }
-
   implicit class OptionError[A](val o: Option[A]) extends AnyVal {
     def getOrError(text: String): A = o.getOrElse {
       throw new NoSuchElementException(text)
