+++ phoenix-scala/test/integration/ImageIntegrationTest.scala
@@ -40,7 +40,7 @@ class ImageIntegrationTest
         val albumResponse = albumsApi(album.formId).get().as[AlbumRoot]
 
         albumResponse.images.length must === (1)
-        albumResponse.images.head.get("src") must not be empty
+        albumResponse.images.head.src must not be empty
       }
 
       "404 if wrong context name" in new Fixture {
